% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_closest_match}
\alias{get_closest_match}
\title{get_closest_match}
\usage{
get_closest_match(species_name, database = NULL, max_dist = 0.15)
}
\arguments{
\item{species_name}{cadena de caracteres (Nombre de genero o nombre científico) la cual quiero buscar en la base de datos.}

\item{database}{La base de datos en la cual quiero ejercer la busqueda, si no está determinada por el usuario, selecciona una interna del paquete.}

\item{max_dist}{Distancia Máxima la cual puede haber entre un elemento de la base de datos y el input para que apareza en el dataframe final.}
}
\value{
DataFrame. Las filas de la base de datos que concuerdan suficientemente con el input
}
\description{
Herramienta de fuzzy match, comparando un input, con una base de datos local
}
\details{
Esta funcion no diferencia entre mayusculas y minusculas, por lo que, es lo mismo "nothofagus", "NOTHOFAGUS" y "nOtHoFaGuS"
Esta funcion es una aplicación de la afind del paquete stringdist. Dentro de la misma, existe una normalización de las distancias para así asegurar un match con hasta 5 letras erroneas. Automaticamente elimina nombres no existentes
}
\examples{
get_closest_match("nothofagus")
Map(get_closest_match(), c("notomircia", "notofagus"))
}
