% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_distribution}
\alias{get_distribution}
\title{get_distribution}
\usage{
get_distribution(species, sinonimos = F)
}
\arguments{
\item{species}{nombre o nombres a buscar.}
}
\value{
Dataframe.El input, el nombre que mejor coincide y las regiones que está presente.
}
\description{
Nos da la prescencia o auscencia de una o mas especies
}
\details{
Cada region se presenta como una columna individual, y su nombre está codificado según los keywords de 3 letras
habituales del Catalogo Taxonómico de la Plantas Vasculares de Chile, publicado en la revista Gayana Botánica (R. Rodriguez, C. Marticorena. et al. 2018)
}
\examples{
get_distribution("Calceolaria")
get_distribution(c("Baccharis linearis", "Mutisia rosea"))
}
