% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_species_by_region}
\alias{get_species_by_region}
\title{get_species_by_regions}
\usage{
get_species_by_region(search, region)
}
\arguments{
\item{search}{str. Grupo o grupos a consultar. Pueden ser: genero, familia, o cualquier taxón superior.}

\item{region}{str. Regiones en las cual buscar. el formato es usando un keyword de 3 letras.}
}
\value{
Lista. Cada lista tiene las especies por grupo y region.
}
\description{
Obtener una lista de especies por region
}
\examples{
get_species_by_region(search="asteraceae", region=c("BIO", "ARA"))

}
