% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_synonyms}
\alias{get_synonyms}
\title{get_synonyms}
\usage{
get_synonyms(name)
}
\arguments{
\item{name}{nombre o nombres a buscar}
}
\value{
Dataframe. La primera fila siempre son los nombres actualizados, (input) y hacia abajo, sus respectivos sinonimos
}
\description{
se obtienenen los sinonimos en base a los nombres actualizados
}
\examples{
get_synonyms(c("Myrceugenia planipes", "Phycella angustifolia"))
get_synonyms("Austrocedrus chilensis")

}
