% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_taxonomy}
\alias{get_taxonomy}
\title{get_taxonomy}
\usage{
get_taxonomy(species_name)
}
\arguments{
\item{species_name}{nombre o nombres a buscar}
}
\value{
Dataframe. Describiendo la taxonomia a detalle
}
\description{
Nos permite recuperar la taxonomía para un nombre, ya sea valido o sinonimo.
}
\examples{
get_taxonomy(c("tupa mucronata", "myrceugenia planipes"))

}
