% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_valid_name}
\alias{get_valid_name}
\title{get_valid_name}
\usage{
get_valid_name(name, use = "user")
}
\arguments{
\item{name}{Nombre o nombres a buscar.}

\item{use}{Formato de input para la funcion. puede ser "user" o "nested". Automaticamente es "user"}
}
\value{
Dataframe. Conteniendo el Input_name y el nombre actualizado
}
\description{
Nos permite obtener el nombre actualizado para uno o varios sinonimos.
}
\details{
En caso de usar nested, el input es el output de la funcion "get_closest_match". La opción "nested" existe para el uso interno de otras funciones dentro de este mismo paquete y no se recomienda al usuario final utilizar esta opcion.
}
\examples{
get_valid_name("myrceugenia")
get_valid_name(c("acacia caven", "tupa mucronata"))
}
