% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_voucher_data}
\alias{get_voucher_data}
\title{get_voucher_data}
\usage{
get_voucher_data(input_name, links = F)
}
\arguments{
\item{input_name}{str. Nombre o nombres a buscar}

\item{links}{Logical. Si links igual a TRUE, se obtienen 2 columnas nuevas, las cuales son los hipervinculos a las imagenes escaladas a diferentes porcentajes.}
}
\value{
Dataframe. 2 columnas el nombre científico y el código de identificación de herbario.
}
\description{
permite obtener el numero total de Vouchers (muestras de herbario) digitalizados para una o mas especies.
}
\details{
Se puede obtener el numero total de vouchers, con su respectivo codigo de herbario (ej. CONC:001, ULS:005)
}
\examples{
get_voucher_data(input_name="Acrisione cymosa", links = T)
}
