% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{porcentajes_fl_cl}
\alias{porcentajes_fl_cl}
\title{porcentajes_fl_cl}
\usage{
porcentajes_fl_cl(
  search = NULL,
  region = NULL,
  by = "family",
  subset = NULL,
  cuenta = FALSE
)
}
\arguments{
\item{search}{string. Taxon o Taxones a buscar}

\item{region}{string. Regiones en las cuales centrar la búsqueda.}

\item{by}{string. Criterio en el cual puedo hacer la separación. Default Familia}

\item{cuenta}{Logical. Agrega una columna al DataFrame final. Si es TRUE, me dice el numero de taxones infraespecifícos}
}
\value{
List. Cada elemento de una lista trae un DataFrame el cual contiene toda la información para el grupo.
}
\description{
Calcula los porcentajes de taxones Nativos, Endemicos, o Introducidos por grupo.
}
\details{
Esta funcion presenta parametros complejos, pero tanto el parametro by,
como el subset pueden ser las siguientes opciones:
·division
·classname
·order
·family
·genus
·plant_habit
·env_habit
·cycle
·conservation_state
El parametro by y el subset NO PUEDEN SER EL MISMO CRITERIO.
}
\examples{
porcentajes_fl_cl(search=c("asteraceae", "apiales"), by="genus", cuenta = T, subset="plant_habit")

}
